backend/requirements.txt
# WasteWise AI — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Rate Limiting ─────────────────────────────────────────────────────────
slowapi==0.1.9

# ── HTTP Client ───────────────────────────────────────────────────────────
httpx==0.27.0

# ── Data Validation ───────────────────────────────────────────────────────
pydantic==2.8.2
python-dotenv==1.0.1

# ── Database ─────────────────────────────────────────────────────────────
supabase==2.4.3

# ── Security ─────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ── Email ─────────────────────────────────────────────────────────────────
resend==2.0.0

# ── AI Providers (Fallback Chain: Gemini → Groq → Mistral) ───────────────
google-generativeai==0.7.2
groq==0.9.0
mistralai==1.0.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Cache & Task Queue (OPTIONAL — for multi-instance/production deployment) ──
# Set REDIS_URL in .env to activate Redis cache instead of in-memory
# Set CELERY_BROKER_URL in .env to activate Celery instead of thread scheduler
redis==5.0.4
celery==5.4.0

# ── File Processing ───────────────────────────────────────────────────────────
pillow==10.3.0
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2

# ── Machine Learning ─────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.0

# ── Text-to-Speech ───────────────────────────────────────────────────────────
edge-tts==6.1.10

# ── Language Detection ───────────────────────────────────────────────────────
langdetect==1.0.9

# ── Web Scraping (for crawler) ────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Image OCR (for menu board scanning) ──────────────────────────────────────
easyocr==1.7.1


